+++ stdnet/conf.py
@@ -27,6 +27,7 @@ To change settings::
     def __init__(self):
         self.DEFAULT_BACKEND = 'redis://127.0.0.1:6379/?db=7'
         self.DEFAULT_KEYPREFIX  = 'stdnet.'
+        self.CHARSET = 'utf-8'
         self.SCHEMA = 'compact-hash'
         self.REDIS_PARSER = None
         self.MAX_CONNECTIONS = 2**31