+++ stdnet/utils/__init__.py
@@ -5,9 +5,11 @@ from stdnet.lib.py2py3.py2py3 import *
 if ispy3k:
     import pickle
     from io import BytesIO
+    chr
 else:
     import cPickle as pickle
-    from cStringIO import StringIO as BytesIO 
+    from cStringIO import StringIO as BytesIO
+    chr = unichr 
     
 
 from .jsontools import *